executor: fix the definition of struct tss64
Per https://wiki.osdev.org/Task_State_Segment#Long_Mode,
io_bitmap and reserved3 should be 16-bit.

+++ executor/common_kvm_amd64.h
@@ -80,8 +80,8 @@ struct tss64 {
 	uint64 reserved1;
 	uint64 ist[7];
 	uint64 reserved2;
-	uint32 reserved3;
-	uint32 io_bitmap;
+	uint16 reserved3;
+	uint16 io_bitmap;
 } __attribute__((packed));
 
 static void fill_segment_descriptor(uint64* dt, uint64* lt, struct kvm_segment* seg)
